% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{api_basic_auth}
\alias{api_basic_auth}
\title{Basic Auth}
\usage{
api_basic_auth(baseurl = "https://clone.psi-mis.org/", username, password)
}
\arguments{
\item{baseurl}{A string, the base URL of a PSI - MIS. Default is PSI - MIS clone server.}

\item{username}{A string. PSI - MIS username.}

\item{password}{A string. PSI - MIS user password.}
}
\value{
logical
}
\description{
Authenticate to a PSI - MIS.
}
