% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{api_endpoint}
\alias{api_endpoint}
\title{Generate an API Endpoint}
\usage{
api_endpoint(
  baseurl = "https://clone.psi-mis.org/",
  dimension_dx = c("dQTWxMDtAiW", "mOGarPwHuFc"),
  category = "n5ODfcdD1YQ",
  category_option = "YF8v2OSxWKl",
  dimension_ou = "rP1W74RpNWF",
  dimension_pe = "THIS_MONTH",
  completed_only = "false"
)
}
\arguments{
\item{baseurl}{A string, the base URL of a PSI - MIS. Default is PSI - MIS
clone server.}

\item{dimension_dx}{A string or character vector, data elements or indicators
uid. The defaults are Oraq (id: mOGarPwHuFc) and Insti (id: dQTWxMDtAiW)
kits bought data elements uids.}

\item{category}{A string, category UID. Default; KE HIVST - Distribution
channels (id: n5ODfcdD1YQ)}

\item{category_option}{A string, the distribution model. Default is hotspot
(id: YF8v2OSxWKl)}

\item{dimension_ou}{A string or character vector, the orgUnit UID. Default is
Kenya (id: rP1W74RpNWF).}

\item{dimension_pe}{A string or character vector, DHIS2 period. The default
is THIS_MONTH}

\item{completed_only}{Logical, whether to pull only the completed events. The
default is false.}
}
\value{
A string, An API endpoint.
}
\description{
Create an API endpoint to query kits from the analytics.
}
\details{
\code{api_endpoint} constructs an API endpoint to query HIV self testing kits
from the analytics resource. The default endpoint will query insti and oraq
kits distributed via hotpots.
}
