% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{api_get}
\alias{api_get}
\title{Pull a Resource at PSI-MIS}
\usage{
api_get(endpoint = NULL)
}
\arguments{
\item{endpoint}{A string, the API endpoint}
}
\value{
An S3 object.
}
\description{
Pull a Resource at PSI-MIS
}
