% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{api_json_check}
\alias{api_json_check}
\title{Check an API response is JSON}
\usage{
api_json_check(resp = NULL)
}
\arguments{
\item{resp}{A response}
}
\value{
An error message
}
\description{
Check that an API request returned a JSON object.
}
