% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{api_resource}
\alias{api_resource}
\title{API resource}
\usage{
api_resource(resource = NULL, push = FALSE)
}
\arguments{
\item{resource}{A string, A PSI - MIS resource.}

\item{push}{Logical. Default is FALSE, which returns the analytics resource.}
}
\value{
A string, an API resource
}
\description{
Point an API resource.
}
\details{
\code{api_resource} sets an API resource.
}
