% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{api_status_check}
\alias{api_status_check}
\title{Check an API response}
\usage{
api_status_check(resp = NULL)
}
\arguments{
\item{resp}{An API response to check.}
}
\value{
An error message
}
\description{
Did an API request return error?
}
