% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{api_update_data}
\alias{api_update_data}
\title{Update Data Values at PSI MIS}
\usage{
api_update_data(
  endpoint = NULL,
  data_values = NULL,
  import_strategy = "CREATE_AND_UPDATE"
)
}
\arguments{
\item{endpoint}{A string, API endpoint.}

\item{data_values}{A data.frame, filtered kits.}

\item{import_strategy}{A string. Default is CREATE_UPDATE.}
}
\value{
An S3 object, reponse
}
\description{
Create or update dataValueSet resource at PSI MIS.
}
