% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{api_version}
\alias{api_version}
\title{API version}
\usage{
api_version(version = NULL)
}
\arguments{
\item{version}{Numeric. The current supported Web API version.}
}
\value{
The supported API version.
}
\description{
Set an API version
}
