% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bypass.R
\name{bypass_ke_distr}
\alias{bypass_ke_distr}
\title{Bypass KE HIVST Distribution channels}
\usage{
bypass_ke_distr(baseurl)
}
\arguments{
\item{baseurl}{A string, the base URL of a PSI - MIS. Default is PSI - MIS clone server.}
}
\description{
Bypass KE Distribution channels, a category combination filter.
}
