% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_status}
\alias{check_status}
\title{Check the status of API request
@importFrom httr http_error status_code
@param resp API response.}
\usage{
check_status(resp)
}
\description{
Check the status of API request
@importFrom httr http_error status_code
@param resp API response.
}
