% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hivstr.R
\docType{package}
\name{hivstr}
\alias{hivstr}
\title{hivstr: A package to bypass HIV self testing distribution channels}
\description{
The hivstr package bypasses HIV Self Testing Distribution Channels, a
category combination filter, directly from the analytic views. It makes it
easier to combine/visualize HIV self testing data with other sources, build
indicators and visualize data on a single chart/table.
}
