% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bypass.R
\name{kits_data_elements}
\alias{kits_data_elements}
\title{Kits data elements}
\usage{
kits_data_elements(type)
}
\arguments{
\item{type}{A string.}
}
\value{
A list
}
\description{
Mapping of new kits data elements.
}
