% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bypass.R
\name{kits_endpoint}
\alias{kits_endpoint}
\title{Kits endpoints}
\usage{
kits_endpoint(baseurl)
}
\arguments{
\item{baseurl}{A string, the base URL of a PSI - MIS. Default is PSI - MIS clone server.}
}
\value{
An S3 object.
}
\description{
Kits endpoints
}
